New Implementation of Spawning a weapon
This is kinda cheeky and will probably get me scalded in some
communities, but I made it when you instantiate a new Instance it takes
everything needed to put a newly generated item in the players
inventory. Keep in mind this has nothing to do with the old Weapon.java
that was deleted a commit earlier.