Added Init Checks for addons
This was cause in the onEnable of every plugin it calls this method. So
the dependables checks if it has been toggled to show its this table.
And the isInit tells us if that plugin has actually been initialized so
we dont try to check if all the tables in subsequent order exist
everytime a new plugin boots.